split sponsors for lazy loading
